[poly] separate polynomial module
