+++ schnorr_fun/Cargo.toml
@@ -44,7 +44,7 @@ serde = ["secp256kfun/serde"]
 libsecp_compat = ["secp256kfun/libsecp_compat", "libsecp_compat_0_27"]
 libsecp_compat_0_27 = ["secp256kfun/libsecp_compat_0_27"]
 proptest = ["secp256kfun/proptest"]
-frost_backup = ["bech32"]
+share_backup = ["bech32"]
 
 [dependencies.bech32]
 version = "0.9"
