+++ schnorr_fun/src/share_backup.rs
@@ -54,10 +54,8 @@ use secp256kfun::{
     g,
     hash::HashAdd,
     marker::{PointType, Public, Secrecy, Zero, ZeroChoice},
-    s, Point, Scalar,
+    poly, s, Point, Scalar,
 };
-
-use crate::frost;
 
 /// An error encountered when encoding a Frostsnap backup.
 #[derive(Debug, Copy, Clone)]
@@ -187,101 +185,4 @@ pub fn decode_backup(
     };
 
     Ok((threshold, identifier, secret_share, share_index))
-}
-
-/// Find the coefficients of the polynomial that interpolates a set of points at unique indexes.
-///
-/// Panics if the indexes are not unique.
-///
-/// A vector with a tail of zero coefficients means the interpolation was overdetermined.
-pub fn interpolate_point_polynomial(
-    indexes: Vec<Scalar<impl Secrecy, impl ZeroChoice>>,
-    points: Vec<Point>,
-) -> Vec<Point<impl PointType, Public, Zero>> {
-    // Get each lagrange basis polynomial from the product of coefficients:
-    //      l_j(x) = Product[ (x-x_m)/(x_j-x_m), j!=m]
-    // Or
-    //      l_j(x) = Product[ a_m*x + b_m, j!=m], where a_m = 1/(x_j-x_m) and b_m = -x_m*a_m.
-    let basis_polynomials: Vec<_> = indexes
-        .clone()
-        .into_iter()
-        .enumerate()
-        .map(|(j, x_j)| {
-            let mut coefficients: Vec<_> = vec![];
-            for (_, x_m) in indexes.iter().enumerate().filter(|(m, _)| *m != j) {
-                let a_m = s!(x_j - x_m)
-                    .non_zero()
-                    .expect("points must lie at unique indexes to interpolate")
-                    .invert();
-                let b_m = s!(-x_m * a_m).mark_zero();
-
-                // Multiply out the product. Beginning with the first two coefficients
-                // we then take the next set (b_1, a_1), multiply through, and collect terms.
-                if coefficients.is_empty() {
-                    coefficients.extend([b_m.mark_zero(), a_m.mark_zero()])
-                } else {
-                    let mut updated_coefficients = coefficients.clone();
-                    for i in 0..coefficients.len() {
-                        let bumping_up_degree = if i != 0 {
-                            s!({ &coefficients[i - 1] } * a_m)
-                        } else {
-                            s!(0)
-                        };
-                        let same_degree = s!({ &coefficients[i] } * b_m);
-                        updated_coefficients[i] = s!(same_degree + bumping_up_degree);
-                    }
-                    let higher_degree = s!({ coefficients.last().unwrap() } * a_m);
-                    updated_coefficients.push(higher_degree);
-                    coefficients = updated_coefficients;
-                }
-            }
-            coefficients
-        })
-        .collect();
-
-    let interpolating_basis: Vec<_> = basis_polynomials
-        .iter()
-        .zip(points)
-        .map(|(basis_poly, y_value)| {
-            basis_poly
-                .iter()
-                .map(|coeff| g!(coeff * y_value))
-                .collect::<Vec<_>>()
-        })
-        .collect();
-
-    let mut point_polynomial = interpolating_basis[0].clone();
-    for poly in &interpolating_basis[1..] {
-        for (i, point) in poly.iter().enumerate() {
-            point_polynomial[i] = g!({ point_polynomial[i] } + point);
-        }
-    }
-
-    point_polynomial
-}
-
-/// Interpolate a set of shamir secret shares to find the joint secret
-///
-/// Panics if the indexes are not unique.
-pub fn reconstruct_shared_secret(
-    indexes: Vec<Scalar>,
-    secrets: Vec<Scalar<impl Secrecy, impl ZeroChoice>>,
-) -> Scalar {
-    let coefficients: Vec<_> = indexes
-        .iter()
-        .map(|my_index| {
-            frost::lagrange_lambda(
-                my_index.clone(),
-                indexes.clone().into_iter().filter(|j| j != my_index),
-            )
-        })
-        .collect();
-    secrets
-        .into_iter()
-        .zip(coefficients)
-        .fold(s!(0).mark_zero(), |acc, (secret, coefficient)| {
-            s!(acc + secret * coefficient)
-        })
-        .non_zero()
-        .expect("joint secret should not be zero")
 }
